Multi-scope fixes for themes and module configurations

+++ core/modules/configuration/Actions.php
@@ -733,6 +733,7 @@
             $this->writable_link = is_writable(THEBUGGENIE_PATH . THEBUGGENIE_PUBLIC_FOLDER_NAME . DS . 'themes');
             $this->theme_message = framework\Context::getMessageAndClear('theme_message');
             $this->theme_error = framework\Context::getMessageAndClear('theme_error');
+            $this->is_default_scope = framework\Context::getScope()->isDefault();
         }
 
         /**
@@ -743,10 +744,12 @@
          */
         public function runUpdateModule(framework\Request $request)
         {
-            $module = framework\Context::getModule($request['module_key']);
-            $module->upgrade();
-            $module->enable();
-            framework\Context::setMessage('module_message', $this->getI18n()->__('The module was updated'));
+            if (framework\Context::getScope()->isDefault()) {
+                $module = framework\Context::getModule($request['module_key']);
+                $module->upgrade();
+                $module->enable();
+                framework\Context::setMessage('module_message', $this->getI18n()->__('The module was updated'));
+            }
             $this->forward($this->getRouting()->generate('configure_modules'));
         }
 
@@ -761,12 +764,15 @@
         {
             $themes = framework\Context::getThemes();
             if (array_key_exists($request['theme_key'], $themes)) {
-                $theme_link_path = THEBUGGENIE_PATH . THEBUGGENIE_PUBLIC_FOLDER_NAME . DS . 'css' . DS . $request['theme_key'];
-                $theme_path = '..' . DS . '..' . DS . 'themes' . DS . $request['theme_key'] . DS . 'css';
-                if (file_exists($theme_link_path)) {
-                    unlink($theme_link_path);
+                if (framework\Context::getScope()->isDefault())
+                {
+                    $theme_link_path = THEBUGGENIE_PATH . THEBUGGENIE_PUBLIC_FOLDER_NAME . DS . 'css' . DS . $request['theme_key'];
+                    $theme_path = '..' . DS . '..' . DS . 'themes' . DS . $request['theme_key'] . DS . 'css';
+                    if (file_exists($theme_link_path)) {
+                        unlink($theme_link_path);
+                    }
+                    symlink($theme_path, $theme_link_path);
                 }
-                symlink($theme_path, $theme_link_path);
                 framework\Settings::saveSetting(framework\Settings::SETTING_THEME_NAME, $request['theme_key']);
                 framework\Context::setMessage('theme_message', $this->getI18n()->__('The theme has been enabled'));
             } else {
